Added Shared Project and eliminated cource files linking
Added Shared project and common logic moved to it eliminating source files linking
Impoved XML comment for AuthenticationBroker.AuthenticateAsync
Added common abstact class AuthenticationException to be able to catch all Auth0 exceptions at once
