+++ Auth0Client.Phone.Shared/AssemblyInfo.cs
@@ -1,0 +1,15 @@
+﻿using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+// General Information about an assembly is controlled through the following
+// set of attributes. Change these attribute values to modify the information
+// associated with an assembly.
+[assembly: AssemblyTitle("Auth0 Client Library")]
+[assembly: AssemblyDescription("")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct("Auth0")]
+[assembly: AssemblyCopyright("Copyright © Auth0 2015")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
