+++ Auth0Client.Phone.Shared/Auth0Client.Phone.Shared.projitems
@@ -1,0 +1,19 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>73bc5d17-03b8-4e99-beea-a284fbfbc27e</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>Auth0Client.Phone.Shared</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)AssemblyInfo.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Auth0Client.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Auth0User.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)AuthenticationBroker.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)AuthenticationExceptions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utils.cs" />
+  </ItemGroup>
+</Project>
