+++ Auth0Client.Phone.Shared/AuthenticationExceptions.cs
@@ -1,0 +1,29 @@
+﻿using System;
+
+namespace Auth0.SDK
+{
+    public abstract class AuthenticationException : Exception
+    {
+        protected AuthenticationException()
+        {
+        }
+
+        protected AuthenticationException(string message) : base(message)
+        {
+        }
+    }
+
+    public class AuthenticationErrorException : AuthenticationException
+    {
+        public AuthenticationErrorException(string message) : base(message)
+        {
+        }
+    }
+
+    public class AuthenticationCancelException : AuthenticationException
+    {
+        public AuthenticationCancelException()
+        {
+        }
+    }
+}
