--- Auth0Client/AuthenticationCancelException.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace Auth0.SDK
-{
-    public class AuthenticationCancelException : Exception
-    {
-        public AuthenticationCancelException()
-        {
-        }
-    }
-}
