--- Auth0Client/AuthenticationErrorException.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace Auth0.SDK
-{
-    public class AuthenticationErrorException : Exception
-    {
-        public AuthenticationErrorException(string message) : base(message)
-        {
-        }
-    }
-}
