+++ SampleApp/MainPage.xaml.cs
@@ -21,7 +21,7 @@ namespace SampleApp
                 "contoso.auth0.com",
                 "cEPMBtnqXQdFBuUyJeAugh6W7kLIoepB");
 
-        private async void Button_Click(object sender, RoutedEventArgs e)
+        private async void Login_Click(object sender, RoutedEventArgs e)
         {
             try
             {
